Added in the ability to click on the CPU usage widget
Adding in the ability to click on this widget I wanted
users to be able to use either 'top' or 'activity monitor'.

Because 'top' is run in a terminal I wanted users to
be able to specify their preferred Terminal.

I added two terminal options Terminal and iTerm2.

Hopefully the ability to open a terminal of the users
choosing and run an arbitrary command will be helpful
else where in the project.